[Data Structure] Queue
Implementation using ListNode
